Update HuntAndPeck.pyw
Oh, my app id, Hunt and Peck, 
How thou art strange and meek. 
A name so simple, yet so grand, 
It brings me Joy to type by hand.
With every key that I press, 
I have to guess less and less.
As my fingers Hunt and Peck upon the keys, 
I find them all as if it's just a breeze.